+++ android/app/src/main/java/com/woowacourse/friendogly/presentation/ui/group/list/GroupListFragment.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.friendogly.presentation.ui.grouplist
+package com.woowacourse.friendogly.presentation.ui.group.list
 
 import com.woowacourse.friendogly.R
 import com.woowacourse.friendogly.databinding.FragmentGroupListBinding
