+++ android/app/src/main/res/layout/fragment_group_list.xml
@@ -4,7 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <data>
-
+        <variable
+            name="vm"
+            type="com.woowacourse.friendogly.presentation.ui.group." />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
