+++ android/app/src/main/res/layout/include_group_filter_selector.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="groupFilterSelector"
-            type="com.woowacourse.friendogly.presentation.ui.grouplist.grouplistUiModel.GroupFilterSelector" />
+            type="com.woowacourse.friendogly.presentation.ui.group.list.model.GroupFilterSelector" />
 
     </data>
